+++ src/Shipment/Contracts/ShipmentEvent.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Shipment\Contracts;
+
+interface ShipmentEvent
+{
+    public function getShipment(): Shipment;
+}