+++ src/Shipment/Events/ShipmentIsReady.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Vanilo\Shipment\Events;
+
+class ShipmentIsReady extends BaseShipmentEvent
+{
+}